+++ src/components/App.styled.jsx
@@ -3,6 +3,7 @@ import styled from '@emotion/styled';
 export const Container = styled.div`
   background-color: #f5f6fb;
   padding: 0 20px;
+  
 
   @media (min-width: 767px) and (max-width: 1279.98px) {
     padding: 0 171px;
